feat[#126] : ItemContent 컴포넌트 분리 - close #126
- ItemContent 컴포넌트를 별도 파일로 분리하였습니다.
